+++ docs/CHANGELOG.rst
@@ -39,6 +39,7 @@ Added
 - Add SnpEff tool and bokeh assets to ``resolwebio/dnaseq`` Docker image
 - Add automated library strand detection to ``feature_counts`` quantification process
 - Add FastQC option ``nogroup`` to ``bbduk-single`` and ``bbduk-paired`` processes
+- Add CPM normalization to ``htseq-count-raw`` process
 
 Fixed
 -----